+++ test/helpers.cpp
@@ -94,6 +94,12 @@ void testRobotStateIsZero(const franka::RobotState& actual) {
   for (double element : actual.K_F_ext_hat_K) {
     EXPECT_EQ(0.0, element);
   }
+  for (double element : actual.theta) {
+    EXPECT_EQ(0.0, element);
+  }
+  for (double element : actual.dtheta) {
+    EXPECT_EQ(0.0, element);
+  }
   EXPECT_FALSE(actual.current_errors);
   EXPECT_FALSE(actual.last_motion_errors);
   EXPECT_EQ(0u, actual.time.toMSec());
@@ -130,6 +136,8 @@ void testRobotStatesAreEqual(const franka::RobotState& expected, const franka::R
   EXPECT_EQ(expected.tau_ext_hat_filtered, actual.tau_ext_hat_filtered);
   EXPECT_EQ(expected.O_F_ext_hat_K, actual.O_F_ext_hat_K);
   EXPECT_EQ(expected.K_F_ext_hat_K, actual.K_F_ext_hat_K);
+  EXPECT_EQ(expected.theta, actual.theta);
+  EXPECT_EQ(expected.dtheta, actual.dtheta);
   EXPECT_EQ(expected.current_errors, actual.current_errors);
   EXPECT_EQ(expected.last_motion_errors, actual.last_motion_errors);
   EXPECT_EQ(expected.control_command_success_rate, actual.control_command_success_rate);
@@ -173,6 +181,8 @@ void testRobotStatesAreEqual(const research_interface::robot::RobotState& expect
   EXPECT_EQ(expected.tau_ext_hat_filtered, actual.tau_ext_hat_filtered);
   EXPECT_EQ(expected.O_F_ext_hat_K, actual.O_F_ext_hat_K);
   EXPECT_EQ(expected.K_F_ext_hat_K, actual.K_F_ext_hat_K);
+  EXPECT_EQ(expected.theta, actual.theta);
+  EXPECT_EQ(expected.dtheta, actual.dtheta);
   EXPECT_EQ(franka::Errors(expected.errors), actual.current_errors);
   EXPECT_EQ(franka::Errors(expected.reflex_reason), actual.last_motion_errors);
   EXPECT_EQ(expected.message_id, actual.time.toMSec());
@@ -295,6 +305,12 @@ void randomRobotState(franka::RobotState& robot_state) {
   for (double& element : robot_state.K_F_ext_hat_K) {
     element = randomDouble();
   }
+  for (double& element : robot_state.theta) {
+    element = randomDouble();
+  }
+  for (double& element : robot_state.dtheta) {
+    element = randomDouble();
+  }
   std::array<bool, sizeof(research_interface::robot::RobotState::errors)> errors{};
   for (bool& error : errors) {
     error = randomBool();
@@ -380,6 +396,12 @@ void randomRobotState(research_interface::robot::RobotState& robot_state) {
     element = randomDouble();
   }
   for (double& element : robot_state.K_F_ext_hat_K) {
+    element = randomDouble();
+  }
+  for (double& element : robot_state.theta) {
+    element = randomDouble();
+  }
+  for (double& element : robot_state.dtheta) {
     element = randomDouble();
   }
   for (bool& error : robot_state.errors) {