+++ test/robot_command_tests.cpp
@@ -8,9 +8,9 @@
 #include "mock_server.h"
 
 using franka::CommandException;
-using franka::ProtocolException;
 using franka::IncompatibleVersionException;
 using franka::Network;
+using franka::ProtocolException;
 using franka::RealtimeConfig;
 using franka::Robot;
 using franka::RobotState;
@@ -25,6 +25,7 @@ using research_interface::robot::SetCartesianImpedance;
 using research_interface::robot::SetCollisionBehavior;
 using research_interface::robot::SetEEToK;
 using research_interface::robot::SetFToEE;
+using research_interface::robot::SetFilters;
 using research_interface::robot::SetGuidingMode;
 using research_interface::robot::SetJointImpedance;
 using research_interface::robot::SetLoad;
@@ -123,6 +124,20 @@ bool Command<SetFToEE>::compare(const SetFToEE::Request& request_one,
 }
 
 template <>
+bool Command<SetFilters>::compare(const SetFilters::Request& request_one,
+                                  const SetFilters::Request& request_two) {
+  return request_one.joint_position_filter_frequency ==
+             request_two.joint_position_filter_frequency &&
+         request_one.joint_velocity_filter_frequency ==
+             request_two.joint_velocity_filter_frequency &&
+         request_one.cartesian_position_filter_frequency ==
+             request_two.cartesian_position_filter_frequency &&
+         request_one.cartesian_velocity_filter_frequency ==
+             request_two.cartesian_velocity_filter_frequency &&
+         request_one.controller_filter_frequency == request_two.controller_filter_frequency;
+}
+
+template <>
 bool Command<SetLoad>::compare(const SetLoad::Request& request_one,
                                const SetLoad::Request& request_two) {
   return request_one.F_x_Cload == request_two.F_x_Cload &&
@@ -202,6 +217,11 @@ SetFToEE::Request Command<SetFToEE>::getExpected() {
 }
 
 template <>
+SetFilters::Request Command<SetFilters>::getExpected() {
+  return SetFilters::Request(1, 10, 100, 100, 1000);
+}
+
+template <>
 SetLoad::Request Command<SetLoad>::getExpected() {
   double m_load = 1.5;
   std::array<double, 3> F_x_Cload{0.01, 0.01, 0.1};
@@ -247,6 +267,7 @@ using CommandTypes = ::testing::Types<GetCartesianLimit,
                                       SetEEToK,
                                       SetFToEE,
                                       SetLoad,
+                                      SetFilters,
                                       Move,
                                       StopMove,
                                       AutomaticErrorRecovery>;
@@ -312,7 +333,8 @@ using SetterCommandTypes = ::testing::Types<SetCollisionBehavior,
                                             SetCartesianImpedance,
                                             SetEEToK,
                                             SetFToEE,
-                                            SetLoad>;
+                                            SetLoad,
+                                            SetFilters>;
 
 TYPED_TEST_CASE(SetterCommand, SetterCommandTypes);