+++ CHANGELOG.md
@@ -8,13 +8,14 @@ Requires Panda system version >= 1.1.0
 
   * Added two new errors to the robot state
   * Added `epsilon` parameters to gripper grasp method
-  * Added `tau_J_d`, `m_ee`, `F_x_Cee`, `I_ee`, `m_total`, `F_x_Ctotal` and `I_total`
-    to the robot state
+  * Added `tau_J_d`, `m_ee`, `F_x_Cee`, `I_ee`, `m_total`, `F_x_Ctotal`, `I_total`,
+    `theta` and `dtheta` to the robot state
   * Added logs to ControlException
   * Added saturation to joint velocity, joint position, and joint impedance examples
   * Added support for commanding elbow positions for Cartesian motions
   * Fail earlier (by throwing exception) if any of the commanded values are NaN or infinity
   * Added support for stiffness frame to model
+  * Added `setFilter` command
 
 ### Bugfixes